requirements.txt
Flask
Flask-SQLAlchemy
pymysql  
psycopg2-binary
flask-cors